refactor: 컴포넌트 분리 및 zod 적용
