+++ src/app/(auth)/find-pw/found/page.tsx
@@ -1,10 +1,13 @@
-import FindPwPage from '@/components/pages/auth/find-auth/FindPwPage';
+import ChangePwForm from '@/components/forms/ChangePwForm';
+import ChangePwHeader from '@/components/pages/auth/find-pw/ChangePwHeader';
+import { Layout } from '@/components/ui/layout';
 import React from 'react';
 
 export default function Page() {
   return (
-    <>
-      <FindPwPage />
-    </>
+    <Layout variant="findPw">
+      <ChangePwHeader />
+      <ChangePwForm />
+    </Layout>
   );
 }
