+++ src/app/(auth)/sign-in/page.tsx
@@ -1,10 +1,11 @@
-import SignInPage from '@/components/pages/auth/sign-in/SignInPage';
+import SignInForm from '@/components/forms/SignInForm';
+import { Layout } from '@/components/ui/layout';
 import React from 'react';
 
 export default function Page() {
   return (
-    <>
-      <SignInPage />
-    </>
+    <Layout variant="signIn">
+      <SignInForm />
+    </Layout>
   );
 }
