+++ src/app/(auth)/sign-up/page.tsx
@@ -1,10 +1,13 @@
-import SignUpPage from '@/components/pages/auth/sign-up/SignUpPage';
+import SignUpForm from '@/components/forms/SignUpForm';
+import SignUpHeader from '@/components/pages/auth/sign-up/SignUpHeader';
+import { Layout } from '@/components/ui/layout';
 import React from 'react';
 
 export default function Page() {
   return (
-    <>
-      <SignUpPage />
-    </>
+    <Layout variant="authentication">
+      <SignUpHeader />
+      <SignUpForm />
+    </Layout>
   );
 }
