--- src/components/forms/ExampleForm.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function LogInForm() {
-  return <></>;
-}
-
-export default LogInForm;
