--- src/components/layouts/ExampleServiceHeader.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function ExampleServiceHeader() {
-  return <></>;
-}
-
-export default ExampleServiceHeader;
