--- src/components/pages/ExamplePage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function ExamplePage() {
-  return <></>;
-}
-
-export default ExamplePage;
