--- src/components/pages/auth/find-auth/FindPwPage.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import ChangePwHeader from '@/components/auth/ChangePwHeader';
-import ChangePwForm from '@/components/forms/ChangePwForm';
-import { Layout } from '@/components/ui/layout';
-
-function FindPwPage() {
-  return (
-    <Layout variant="findPw">
-      <ChangePwHeader />
-      <ChangePwForm />
-    </Layout>
-  );
-}
-
-export default FindPwPage;
