+++ src/components/pages/auth/sign-up/SignUpIntroHeader.tsx
