+++ src/components/ui/button.tsx
@@ -4,27 +4,27 @@ import { cva, type VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/lib/utils';
 
-const buttonVariants = cva(
-  'inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50',
-  {
-    variants: {
-      variant: {
-        default: 'bg-primary text-primary-foreground hover:bg-primary/90',
-        link: 'text-primary underline-offset-4 hover:underline',
-      },
-      size: {
-        default: 'h-10 px-4 py-2',
-        submit:
-          'h-10 px-4 py-2 w-full bg-green-500 rounded-full disabled:bg-slate-300',
-        kakao: 'w-full h-10 bg-white relative',
-      },
+const buttonVariants = cva('', {
+  variants: {
+    variant: {
+      default: 'bg-primary text-primary-foreground hover:bg-primary/90',
+      link: 'text-primary underline-offset-4 hover:underline',
+      clear:
+        'absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-500 text-xl',
     },
-    defaultVariants: {
-      variant: 'default',
-      size: 'default',
+    size: {
+      default: 'h-10 px-4 py-2',
+      submit:
+        'h-10 px-4 py-2 w-full bg-green-500 rounded-full disabled:bg-slate-300',
+      kakao: 'w-full h-10 bg-white relative',
+      clear: '',
     },
-  }
-);
+  },
+  defaultVariants: {
+    variant: 'default',
+    size: 'default',
+  },
+});
 
 export interface ButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement>,
