+++ src/components/ui/layout.tsx
@@ -13,7 +13,7 @@ const divVariants = cva('', {
       findId: 'flex flex-col h-screen justify-center items-center',
       findPw: 'flex flex-col p-6 mt-10',
       submitDiv:
-        'fixed bottom-0 left-0 right-0 px-4 py-4 disabled:bg-slate-300 border-t-2',
+        'fixed bottom-0 left-0 right-0 px-4 py-4 disabled:bg-slate-300 border-t-2 bg-white',
       main: 'w-full',
     },
   },
