+++ src/types/RequestTypes.ts
@@ -1,7 +1,12 @@
 export interface SignUpRequestType {
   id: string;
+  nickname: string;
   email: string;
   password: string;
   confirmPassword: string;
-  nickname: string;
+}
+
+export interface ChangePwRequestType {
+  password: string;
+  confirmPassword: string;
 }
